Added user auth details in CharacterService

+++ Dotnet-RPG/Services/CharacterService/CharacterService.cs
@@ -4,6 +4,7 @@ using Dotnet_RPG.Dtos.Character;
 using Dotnet_RPG.Models;
 using Microsoft.EntityFrameworkCore;
 using System.Collections.Generic;
+using System.Security.Claims;
 
 namespace Dotnet_RPG.Services.CharacterService
 {
@@ -11,11 +12,19 @@ namespace Dotnet_RPG.Services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             this._mapper = mapper;
             this._context = context;
+            this._httpContextAccessor = httpContextAccessor;
+        }
+
+        private int GetUserId()
+        {
+            return int.Parse(_httpContextAccessor.HttpContext!.User
+            .FindFirstValue(ClaimTypes.NameIdentifier)!);
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
@@ -23,17 +32,24 @@ namespace Dotnet_RPG.Services.CharacterService
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             var character = _mapper.Map<Character>(newCharacter);
 
+            // Assign character and user to each other
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
+            // Add character to DB and save changes
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
 
-            serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            // Return all users current characters
+            serviceResponse.Data = await _context.Characters
+                .Where(c => c.User!.Id == GetUserId())
+                .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int uid)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.Where(c => c.User!.Id == uid).ToListAsync();
+            var dbCharacters = await _context.Characters.Where(c => c.User!.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
             return serviceResponse;
         }
@@ -41,7 +57,8 @@ namespace Dotnet_RPG.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetChracterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User!.Id == GetUserId());
+
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -52,14 +69,17 @@ namespace Dotnet_RPG.Services.CharacterService
 
             try
             {
-                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
+                var dbCharacter = await _context.Characters
+                    .Include(c => c.User)
+                    .FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
 
-                if (dbCharacter == null)
+                // Check that the character exists and the the associated user is the logged in user
+                if (dbCharacter == null || dbCharacter.User!.Id != GetUserId())
                 {
                     throw new NullReferenceException($"Character with Id '{updateCharacter.Id}' not found.");
                 }
 
-                _mapper.Map<Character>(updateCharacter);
+                _mapper.Map(updateCharacter, dbCharacter);
 
                 await _context.SaveChangesAsync();
                 serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
@@ -79,7 +99,7 @@ namespace Dotnet_RPG.Services.CharacterService
             
             try
             {
-                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+                var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User!.Id == GetUserId());
                 
                 if (dbCharacter is null)
                 {
@@ -88,7 +108,9 @@ namespace Dotnet_RPG.Services.CharacterService
 
                 _context.Characters.Remove(dbCharacter);
 
-                serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+                serviceResponse.Data = await _context.Characters
+                    .Where(c =>  c.User!.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
             }
             catch(NullReferenceException ex)
             {
